backend/requirements.txt
# MemeQubit Backend - Python dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web3 & blockchain
web3>=6.14.0
eth-account>=0.10.0

# Quantum simulation (classical simulators for PoC)
# neal is provided by dwave-neal (import neal)
dimod>=0.12.0
dwave-neal>=0.5.0
networkx>=3.2.0
numpy>=1.26.0

# Optional: Qiskit for gate-based simulation
# qiskit>=1.0.0
# qiskit-aer>=0.13.0

# Data & async
redis>=5.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# HTTP client
httpx>=0.26.0

# Monitoring (optional)
# prometheus-client>=0.19.0
